+++ 3-period/java-subway-path-precourse-main/src/test/java/subway/jgrapht/JGraphtTest.java
@@ -1,4 +1,4 @@
-package subway;
+package subway.jgrapht;
 
 import static org.assertj.core.api.Assertions.assertThat;
